Complete rewrite, POSIX, STDIN/OUT, no SED
There is no longer a requirement on GNU SED, the file uses POSIX read
commands and inputs the file/stdin at the bottom of the while loop. The
number of options/flags has been reduced drastically, hence the arg
parser was rewritten manually with cases instead of with getopts, much
nicer. As now -o output-file is written, a lot of the file checks were
not needed. It just generally turns out that sed was making things way
too annoying with variables etc. and this is MUCH BETTER!!